+++ src/main/java/me/srrapero720/waterframes/common/block/entity/DisplayTile.java
@@ -36,7 +36,7 @@ import static me.srrapero720.waterframes.WaterFrames.LOGGER;
 
 @Mod.EventBusSubscriber(modid = WaterFrames.ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class DisplayTile extends BlockEntity {
-    private static long lagTickTime;
+    private static int lagTickTime;
 
     public final DisplayData data;
     public final DisplayCaps caps;
@@ -56,7 +56,7 @@ public class DisplayTile extends BlockEntity {
 
     public static void setLagTickTime(long ltt) {
         if (ltt < 60000) {
-            lagTickTime = ltt / 50L;
+            lagTickTime = (int) (ltt / 50);
         } else {
             LOGGER.warn("Rejected tick correction of {}ms, overpass watchdog time", ltt);
         }
@@ -229,7 +229,7 @@ public class DisplayTile extends BlockEntity {
         else            DisplayNetwork.sendClient(new TimePacket(this.getBlockPos(), Math.max(data.tick - (5000 / 50), 0), this.data.tickMax, true), this);
     }
 
-    public void syncTime(boolean clientSide, long tick, long maxTick) {
+    public void syncTime(boolean clientSide, int tick, int maxTick) {
         if (clientSide) DisplayNetwork.sendServer(new TimePacket(this.getBlockPos(), tick, maxTick, true));
         else            DisplayNetwork.sendClient(new TimePacket(this.getBlockPos(), tick, maxTick, true), this);
     }
@@ -246,7 +246,7 @@ public class DisplayTile extends BlockEntity {
             if (this.data.tick < this.data.tickMax) {
                 this.data.tick++;
                 if (lagTickTime != 0 && this.isServer()) {
-                    long ticks = this.data.tick + lagTickTime;
+                    int ticks = this.data.tick + lagTickTime;
                     while (ticks > this.data.tickMax) {
                         ticks -= this.data.tickMax;
                     }
